Gestion des trames de service
Utilisation d'objets plutot que de types primitifs
Modification du service de décodage des trames
Amélioration de la présentation des données sur l'interface (on
n'affiche plus les données de latitude / longitude si tout est à 0)

+++ src/groovy/trackr/SigFoxWSData.groovy
@@ -10,8 +10,8 @@ class SigFoxWSData {
     Date time
     Long epochTime
     Boolean duplicate
-    Float signal
-    Float avgSignal
+    Double signal
+    Double avgSignal
     Point position
-    Float rssi
+    Double rssi
 }
